refactoring

+++ firmware/src/main/GSwifi.h
@@ -126,7 +126,7 @@ public:
      * @retval 0 success
      * @retval -1 failure
      */
-    int join (GSSECURITY sec, const char *ssid, const char *pass, int dhcp = 1, char *name = NULL);
+    int8_t join (GSSECURITY sec, const char *ssid, const char *pass, int dhcp = 1, char *name = NULL);
     bool isJoined ();
 
     int listen (uint16_t port);
